move cart calculation methods from ShoppingCart context to CartView
