+++ lib/eye2eye/shopping_cart.ex
@@ -35,19 +35,19 @@ defmodule Eye2eye.ShoppingCart do
 
   def reload_cart(%Cart{} = cart), do: get_cart_by_user_uuid(cart.user_uuid)
 
-  def total_cart_items(%Cart{} = cart) do
-    Enum.reduce(cart.items, 0, fn item, acc ->
-      item.quantity + acc
-    end)
-  end
+  # def total_cart_items(%Cart{} = cart) do
+  #   Enum.reduce(cart.items, 0, fn item, acc ->
+  #     item.quantity + acc
+  #   end)
+  # end
 
-  def total_cart_price(%Cart{} = cart) do
-    Enum.reduce(cart.items, @min_cart_item_price, fn item, acc ->
-      item
-      |> total_item_price()
-      |> Decimal.add(acc)
-    end)
-  end
+  # def total_cart_price(%Cart{} = cart) do
+  #   Enum.reduce(cart.items, @min_cart_item_price, fn item, acc ->
+  #     item
+  #     |> total_item_price()
+  #     |> Decimal.add(acc)
+  #   end)
+  # end
 
   def get_cart_item!(id) do
     Repo.one(
@@ -70,9 +70,9 @@ defmodule Eye2eye.ShoppingCart do
     )
   end
 
-  def total_item_price(%CartItem{} = item) do
-    Decimal.mult(item.product.price, item.quantity)
-  end
+  # def total_item_price(%CartItem{} = item) do
+  #   Decimal.mult(item.product.price, item.quantity)
+  # end
 
   def update_cart_item(%CartItem{} = cart_item, cart_item_attrs) do
     update_attrs = reduce_item_quantity_by_one(cart_item_attrs)
